Implement basic server listening setup for later
This code is meant for me to test on localhost, to see if I can generate public and private keys successfully. I might re-do this in C because I programmed this in the Raspberry Pi, and it was fairly laggy while I was running it (and before), and I want to see if I can speed things up and make it less computationally expensive if I wrote it in a lower-level language like C. Maybe... For now, though it's in Python and the socket cannot connect to a real user since we're running on localhost.