Added YOLO model used during component testing
This YOLOv8s replicated model is slightly different from the actual model, made to be a smaller-scale, easier to train version. The model performed well, and was fairly good at determining whether something was a man or woman. This is good for future datasets --> Deer are much more distinct from humans.